+++ pkg.nix
@@ -1,6 +1,6 @@
 { mkDerivation, base, binary, bytestring, ghc-prim
 , lens-family-core, lib, pipes, pipes-bytestring, pipes-parse
-, smallcheck, tasty, tasty-hunit, tasty-smallcheck, transformers
+, tasty, tasty-hunit, tasty-quickcheck, transformers
 }:
 mkDerivation {
   pname = "pipes-binary";
@@ -12,7 +12,7 @@ mkDerivation {
   ];
   testHaskellDepends = [
     base binary bytestring ghc-prim lens-family-core pipes pipes-parse
-    smallcheck tasty tasty-hunit tasty-smallcheck transformers
+    tasty tasty-hunit tasty-quickcheck transformers
   ];
   homepage = "https://github.com/k0001/pipes-binary";
   description = "Encode and decode binary streams using the pipes and binary libraries";