+++ nixpkgs.nix
@@ -1,4 +1,0 @@
-builtins.fetchTarball {
-  url = "https://github.com/NixOS/nixpkgs/archive/a165aeceda9f9741d15bc2488425daeb06c0707e.tar.gz";
-  sha256 = "0bshzligqwbjlvx4jygv6k4c54mjdvj1d7xmfzgjl6hhybh3wjpd";
-}